pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "feishu-miqroera-mcp"
version = "0.1.0"
description = "Feishu MCP Server - Feishu group agent toolkit (messages, calendar, tasks, cloud docs)"
requires-python = ">=3.11"

dependencies = [
    "mcp[cli]>=1.0.0",
    "lark-oapi>=1.3.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[project.scripts]
feishu-mcp = "feishu_mcp.server:main"
feishu-mcp-listen = "feishu_mcp.webhook.longconn:start_listener"

[tool.setuptools.packages.find]
where = ["src"]
